ref(api): Move getVerifiedAccounts to api
